+++ MobileSDK/Sources/MobileSDK/Constants.swift
@@ -53,6 +53,8 @@ struct Constants {
         }
 
         switch environment {
+        case .production: return "v1.117.0"
+        case .sandbox, .staging: return "v1.117.0"
         case .production: return "v1.113.2"
         case .sandbox, .staging: return "v1.113.2"
         }