+++ MobileSDK/Tests/MobileSDKTests/Network/Services/Card/CardServiceMock.swift
@@ -10,7 +10,7 @@ import XCTest
 @testable import MobileSDK
 
 class CardServiceMock: Mockable, CardService {
-    
+
     func createToken(tokeniseCardDetailsReq: TokeniseCardDetailsReq, accessToken: String) async throws -> String {
         let cardTokenRes = loadJSON(filename: "card_tokenisation_success_response", type: CardTokenRes.self)
         return cardTokenRes.resource.data
@@ -19,5 +19,10 @@ class CardServiceMock: Mockable, CardService {
     func createGiftCardToken(tokeniseGiftCardReq: TokeniseGiftCardReq, accessToken: String) async throws -> String {
         let cardTokenRes = loadJSON(filename: "card_tokenisation_success_response", type: CardTokenRes.self)
         return cardTokenRes.resource.data
+    }
+    
+    func getCardBinSchema(accessToken: String) async throws -> [BinSchemaRes.BinSchema] {
+        let schemes = loadJSON(filename: "card_schemes", type: BinSchemaRes.self)
+        return schemes.cardSchemas
     }
 }